requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
discord.py==1.6.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
networkx==2.5
NudeNet==2.0.9
numpy==1.19.5
onnxruntime==1.6.0
opencv-python-headless==4.5.1.48
Pillow==8.1.0
progressbar2==3.53.1
protobuf==3.15.2
pydload==1.0.9
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
